fix(systemConfiguration): Prevent go pointer-to-pointer detection
The GO to CGO pointer tracing code in the runtime was throwing panics when the
callback function was a receiver of a struct due to the pointer being in the heap.

Added some masking to allow the pointer to be passed without being detected
by the runtime library, this isn't really unsafe since the pointer isn't going
to be used by CGO code.