module: remove all usages of github.com/pkg/errors
The module has been updated to remove all usages of
`github.com/pkg/errors` in favour of the built in `errors` package along
with `fmt.Errorf`'s ability to produced wrapped errors.

The minimum Go version of this module has been updated to 1.13 as that
was when the `%w` verb was introduced to the format package.